actualizacion

+++ Topicos de Programacion/Bibliotecas/string.h
@@ -1,6 +1,10 @@
 #ifndef STRING_H_INCLUDED
 #define STRING_H_INCLUDED
-int my_strlen(const char* str);
+
+#include <stdio.h>
+#include <stdlib.h>
+
+size_t my_strlen(const char* str);
 char* my_strcat(char* str1, const char*str2);
 int my_strncmp(const char* str1, const char* str2, int n);
 int countOcurrences(const char* text, const char* word);
